+++ packages/devtools_shared/CHANGELOG.md
@@ -4,8 +4,8 @@ Use of this source code is governed by a BSD-style license that can be
 found in the LICENSE file or at https://developers.google.com/open-source/licenses/bsd.
 -->
 # 13.0.0-wip
-* Deprecated `DevToolsStoreKeys.analyticsEnabled` since this is only used for legacy analytics.
-* **Breaking change:** Removed legacy analytics storage key `DevToolsStoreKeys.isFirstRun`.
+* Deprecated `DevToolsStoreKeys.analyticsEnabled` and `DevToolsStoreKeys.isFirstRun`
+since these were only used for legacy analytics.
 * **Breaking change:** Removed legacy analytics APIs and state cleanup
 (e.g. `apiGetFlutterGAEnabled`, `apiGetDevToolsEnabled`, `apiSetDevToolsEnabled`, `apiGetDevToolsFirstRun`, `apiResetDevTools`).
 * **Breaking change:** Removed public constant `devToolsEnabledPropertyName`.